Support getting the functionID from a .so plugin -> Closes #34
Pass get_configuration_parameter down to the plugin
-> Closes #36

Make sure that virtual DSIDs get the EPCGlobal prefix
in their DSID

+++ core/sim/dsidsim.cpp
@@ -102,7 +102,7 @@ namespace dss {
     return static_cast<double>(m_CurrentValue);
   } // GetValue
 
-  uint8_t DSIDSim::GetFunctionID() {
+  uint16_t DSIDSim::GetFunctionID() {
     return FunctionIDDevice;
   } // GetFunctionID
